requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.1
djangorestframework==3.8.2
idna==2.7
Pillow==5.2.0
psycopg2-binary==2.7.5
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
git+https://ivanmiv@bitbucket.org/ivanmiv/django-microservices.git